# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Neon Murer CMS API
  description: |
    Vollständige REST API für das Neon Murer Content Management System.
    
    ## Authentifizierung
    Die meisten Endpoints erfordern JWT-Token Authentication über den `Authorization: Bearer <token>` Header.
    
    ## Fehlerbehandlung
    Alle Endpoints verwenden einheitliche HTTP-Status-Codes und JSON-Antworten.
    
    ## Rate Limiting
    API-Aufrufe sind auf 100 Requests pro Minute pro IP-Adresse begrenzt.
    
  version: 1.0.0
  contact:
    name: Neon Murer AG
    url: https://neonmurer.ch
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://cms.neonmurer.ch/api  
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /settings/backup/create:
    post:
      tags: [Settings]
      summary: Backup erstellen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Backup erfolgreich erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  filename:
                    type: string
                  size:
                    type: integer
  /settings/backup/list:
    get:
      tags: [Settings]
      summary: Verfügbare Backups auflisten
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Backup-Liste erfolgreich abgerufen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  backups:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                        size:
                          type: integer
                        date:
                          type: string
                          format: date-time
                          
  # HEALTH ENDPOINTS
...